% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coexpression.adjacency}
\alias{coexpression.adjacency}
\title{Adjacency matrix}
\format{
A data frame with 794 rows and 794 variables:
\describe{
  \item{lncRNA}{lncRNA symbol}
  \item{lncRNA}{lncRNA symbol}
  ...
}
}
\source{
\doi{10.1111/jcmm.14458}
}
\usage{
coexpression.adjacency
}
\description{
The adjacency matrix of a co-expression network of lncRNAs and
mRNAs in lung adenocarcinoma that was generated using igraph functions
}
\keyword{datasets}
