% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{double.cent.assess}
\alias{double.cent.assess}
\alias{DCA}
\title{Assessment of innate features and associations of two network centrality measures (dependent and independent)}
\usage{
double.cent.assess(data, nodes.colname, dependent.colname, independent.colname)
}
\arguments{
\item{data}{A data frame contaning the values of two continuous variables and the name of
observations (nodes).}

\item{nodes.colname}{The character format (quoted) name of the column containing
the name of observations (nodes).}

\item{dependent.colname}{The character format (quoted) name of the column containing
the values of the dependent variable.}

\item{independent.colname}{The character format (quoted) name of the column containing
the values of the independent variable.}
}
\description{
This function assesses innate features and the association of two centrality measures
(or any two other continuous variables) from the aspect of distribution mode, dependence,
linearity, monotonicity, partial-moments based correlation, and conditional probability of
deviating from corresponding means in oposite direction. This function assumes one
variable as dependent and the other as independent for regression analyses. The non-linear nature of
the association of two centrality measures is evaluated based on generalized additive models (GAM).
The monotonicity of the association is evaluated based on comparing the squared coefficient of
Spearman correlation and R-squared of rank regression analysis.
Also, the correlation between two variables is assessed via non-linear non-parametric statristics (NNS).
For the conditional probability assessment, the idependent variable is considered as the condition variable.
}
\examples{
MyData <- centrality.measures
My.metrics.assessment <- double.cent.assess(data = MyData,
                                            nodes.colname = "name",
                                            dependent.colname = "BetweennessCentrality",
                                            independent.colname = "NeighborhoodConnectivity")
}
\seealso{
\code{\link[nortest]{ad.test}} for Anderson-Darling test for normality,
\code{\link[mgcv]{gam}} for Generalized additive models with integrated smoothness estimation,
\code{\link[stats]{lm}} for Fitting Linear Models,
\code{\link[Hmisc]{hoeffd}} for Matrix of Hoeffding's D Statistics, and
\code{\link[NNS]{NNS.dep}} for NNS Dependence

Other centrality association assessment functions: 
\code{\link{cond.prob.analysis}()},
\code{\link{double.cent.assess.noRegression}()}
}
\concept{centrality association assessment functions}
\keyword{association_assessment}
\keyword{dependence_assessment}
