% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{graph_from_data_frame}
\alias{graph_from_adjacency_matrix}
\alias{graph_from_incidence_matrix}
\alias{V}
\alias{degree}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{igraph}{\code{\link[igraph]{V}}, \code{\link[igraph]{degree}}, \code{\link[igraph]{graph_from_adjacency_matrix}}, \code{\link[igraph]{graph_from_data_frame}}, \code{\link[igraph]{graph_from_incidence_matrix}}}
}}

