% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{runShinyApp}
\alias{runShinyApp}
\title{Run shiny app}
\usage{
runShinyApp(shinyApp)
}
\arguments{
\item{shinyApp}{The name of the shiny app you want to run. You can get the exact name of the available
shiny apps via the following command.
\emph{list.files(system.file("ShinyApps", package = "influential"))}. Please also note this function is
case-sensitive.}
}
\value{
A shiny app.
}
\description{
Run shiny apps included in the influential R package.
Also, a web-based \href{https://influential.erc.monash.edu/}{Influential Software Package} with a convenient
user-interface (UI) has been developed for the comfort of all users including those without a coding background.
}
\examples{
\dontrun{
runShinyApp(shinyApp = "IVI")
}
}
\keyword{runShinyApp}
