% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{V}
\alias{V}
\alias{vertices}
\title{Vertices of an igraph graph}
\usage{
V(graph)
}
\arguments{
\item{graph}{The graph (an igraph graph)}
}
\description{
This function and all of its descriptions have been obtained from the igraph package.
}
\examples{
MyData <- coexpression.data
My_graph <- graph_from_data_frame(MyData)
My_graph_vertices <- V(My_graph)
}
\seealso{
\code{\link[igraph]{V}} for a complete description on this function
}
\keyword{graph_vertices}
