% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{centrality.measures}
\alias{centrality.measures}
\title{Centrality measures dataset}
\format{
A data frame with 794 rows and 6 variables:
\describe{
  \
  \item{DC}{Degree Centrality}
  \item{CR}{ClusterRank}
  \item{NC}{Neighborhood Connectivity}
  \item{LH_index}{Local H-index}
  \item{BC}{Betweenness Centrality}
  \item{CI}{Collective Influence}
  ...
}
}
\source{
\url{https://doi.org/10.1111/jcmm.14458/}
}
\usage{
centrality.measures
}
\description{
The centrality measures of a co-expression network of lncRNAs and
mRNAs in lung adenocarcinoma
}
\keyword{datasets}
