% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{centrality.measures}
\alias{centrality.measures}
\title{Centrality measures dataset}
\format{A data frame with 794 rows and 4 variables:
\describe{
  \item{BetweennessCentrality}{Betweenness Centrality}
  \item{Degree}{Degree Centrality}
  \item{name}{Node (gene) name}
  \item{NeighborhoodConnectivity}{Neighborhood Connectivity}
  ...
}}
\source{
\url{https://www.ncbi.nlm.nih.gov/pubmed/31211495/}
}
\usage{
centrality.measures
}
\description{
The centrality measures of a co-expression network of lncRNAs and
mRNAs in lung adenocarcinoma
}
\keyword{datasets}
