% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coexpression.data}
\alias{coexpression.data}
\title{Co-expression dataset}
\format{
A data frame with 2410 rows and 2 variables:
\describe{
  \item{lncRNA}{lncRNA symbol}
  \item{Coexpressed.Gene}{Co-expressed gene symbol}
  ...
}
}
\source{
\url{https://doi.org/10.1111/jcmm.14458}
}
\usage{
coexpression.data
}
\description{
A co-expression dataset of lncRNAs and mRNAs in lung adenocarcinoma
}
\keyword{datasets}
